requirements.txt
newsapi-python==0.2.7
openai==1.55.3
readabilipy==0.2.0
slack-sdk==3.33.5
